feat(message-dropdown): extraction into component
