+++ src/views/Home.vue
@@ -1,66 +1,11 @@
 <script setup lang="ts" name="home">
-import { computed, ref } from 'vue';
-import type { TabsPaneContext } from 'element-plus';
+import { computed } from 'vue';
 import { RouterView, useRoute, useRouter } from 'vue-router';
 import SideBar from '../components/SideBar.vue';
+import MessageDropdown from '../components/MessageDropdown.vue';
 
 const route = useRoute();
 const router = useRouter();
-
-const notices = [
-  {
-    logo: '../assets/user.svg',
-    content: '你收到了 14 份新周报',
-    time: '6年前',
-  },
-  {
-    logo: '../assets/user.svg',
-    content: '你收到了 14 份新周报',
-    time: '6年前',
-  },
-  {
-    logo: '../assets/user.svg',
-    content: '你收到了 14 份新周报',
-    time: '6年前',
-  },
-];
-
-const messages = [
-  {
-    type: 'comment',
-    username: '王小虎',
-    content: '这是一条评论',
-    avatar:
-      'https://wpimg.wallstcn.com/f778738c-e4f8-4870-b634-56703b4acafe.gif',
-    time: '6年前',
-  },
-  {
-    type: 'reply',
-    username: '王小龙',
-    content: '这是一条回复',
-    avatar:
-      'https://wpimg.wallstcn.com/f778738c-e4f8-4870-b634-56703b4acafe.gif',
-    time: '6年前',
-  },
-];
-
-const todos = [
-  {
-    title: '任务一',
-    status: '未开始',
-    descript: '这是一条任务描述',
-  },
-  {
-    title: '任务二',
-    status: '进行中',
-    descript: '这是一条任务描述',
-  },
-  {
-    title: '任务三',
-    status: '已完成',
-    descript: '这是一条任务描述',
-  },
-];
 
 const handleCommand = (command: string) => {
   if (command == 'loginout') {
@@ -70,12 +15,6 @@ const handleCommand = (command: string) => {
   } else if (command == 'user') {
     router.push('/user');
   }
-};
-
-const activeName = ref('first');
-
-const handleClickMessage = (tab: TabsPaneContext, event: Event) => {
-  console.log(tab, event);
 };
 
 const breadcrumb = computed(() => {
@@ -106,82 +45,7 @@ const username = localStorage.getItem('username') || '';
           />
         </a>
         <!-- 消息通知 -->
-        <el-dropdown trigger="click">
-          <el-badge :value="12" class="item message">
-            <el-icon size="20px" @click="handleClickMessage">
-              <Bell />
-            </el-icon>
-          </el-badge>
-          <template #dropdown>
-            <el-dropdown-menu>
-              <el-tabs
-                style="width: 300px"
-                v-model="activeName"
-                class="demo-tabs"
-                stretch
-                @tab-click="handleClickMessage"
-              >
-                <el-tab-pane label="通知" name="first"></el-tab-pane>
-                <el-tab-pane label="消息" name="second"> </el-tab-pane>
-                <el-tab-pane label="待办" name="third"> </el-tab-pane>
-              </el-tabs>
-              <div class="dropdown-wrapper" v-if="activeName === 'first'">
-                <div
-                  class="notice-container"
-                  v-for="(notice, index) in notices"
-                  :key="index"
-                >
-                  <div class="notice-logo">
-                    <img :src="notice.logo" alt="" />
-                  </div>
-                  <div class="notice-text">
-                    <div class="notice-content">{{ notice.content }}</div>
-                    <div class="notice-time">{{ notice.time }}</div>
-                  </div>
-                </div>
-              </div>
-              <div class="dropdown-wrapper" v-else-if="activeName === 'second'">
-                <div
-                  class="message-container"
-                  v-for="(message, index) in messages"
-                  :key="index"
-                >
-                  <div class="message-logo">
-                    <img :src="message.avatar" alt="" />
-                  </div>
-                  <div class="message-text">
-                    <div>
-                      {{ message.username }}
-                      {{ message.type === 'comment' ? '评论了你' : '回复了你' }}
-                    </div>
-                    <div class="message-content">{{ message.content }}</div>
-                    <div class="message-time">{{ message.time }}</div>
-                  </div>
-                </div>
-              </div>
-              <div class="dropdown-wrapper" v-else>
-                <div
-                  class="todo-container"
-                  v-for="(todo, index) in todos"
-                  :key="index"
-                >
-                  <div class="todo-title">{{ todo.title }}</div>
-                  <el-tag class="todo-status">{{ todo.status }}</el-tag>
-                  <div class="todo-descript">{{ todo.descript }}</div>
-                </div>
-              </div>
-
-              <el-button-group style="inline-size: 100%">
-                <el-button type="text" style="inline-size: 50%"
-                  >清空通知</el-button
-                >
-                <el-button type="text" style="inline-size: 50%"
-                  >查看更多</el-button
-                >
-              </el-button-group>
-            </el-dropdown-menu>
-          </template>
-        </el-dropdown>
+        <MessageDropdown />
         <!-- 用户名下拉菜单 -->
         <el-dropdown class="user-name" trigger="click" @command="handleCommand">
           <span class="el-dropdown-link">
@@ -239,11 +103,6 @@ const username = localStorage.getItem('username') || '';
     display: flex;
     align-items: center;
 
-    .message {
-      margin: 0px 15px;
-      cursor: pointer;
-    }
-
     .name {
       font-size: 16px;
       padding-left: 8px;
@@ -269,120 +128,5 @@ img {
   vertical-align: middle;
   width: 42px;
   border-radius: 50%;
-}
-
-.notice-container {
-  display: flex;
-  align-items: center;
-  padding: 8px 12px;
-  cursor: pointer;
-
-  &:hover {
-    background-color: #f5f7fa;
-  }
-
-  .notice-logo {
-    width: 32px;
-    height: 32px;
-    border-radius: 50%;
-    overflow: hidden;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: #f5f7fa;
-  }
-
-  .notice-text {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    padding-left: 8px;
-
-    .notice-content {
-      font-size: 14px;
-      color: #606266;
-    }
-
-    .notice-time {
-      font-size: 12px;
-      color: #909399;
-    }
-  }
-}
-
-.message-container {
-  display: flex;
-  align-items: center;
-  padding: 8px 12px;
-  cursor: pointer;
-
-  &:hover {
-    background-color: #f5f7fa;
-  }
-
-  .message-logo {
-    width: 32px;
-    height: 32px;
-    border-radius: 50%;
-    overflow: hidden;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: #f5f7fa;
-  }
-
-  .message-text {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    padding-left: 8px;
-
-    .message-content {
-      font-size: 14px;
-      color: #606266;
-    }
-
-    .message-time {
-      font-size: 12px;
-      color: #909399;
-    }
-  }
-}
-
-.todo-container {
-  display: grid;
-  grid-template:
-    'title status'
-    'descript descript';
-
-  padding: 8px 12px;
-  cursor: pointer;
-
-  &:hover {
-    background-color: #f5f7fa;
-  }
-
-  .todo-title {
-    grid-area: title;
-    font-size: 14px;
-    color: #606266;
-  }
-
-  .todo-status {
-    grid-area: status;
-    font-size: 14px;
-    color: #606266;
-    justify-self: end;
-  }
-
-  .todo-descript {
-    grid-area: descript;
-    font-size: 14px;
-    color: #606266;
-  }
-}
-
-.dropdown-wrapper {
-  padding-block-end: 12px;
 }
 </style>
